adding XSDInfo program, updating testset
I'm adding my XSDInfo program in the current (unfinished) state.
It reads the XSD file and produces some output, but it's not working properly.
I'll update it when I have time for that.

I've also created another file in the testset folder called index.xml.
This file contains a 'kind of' scheme for the xml structure.
Maybe that's a help for manually creating a documentary.
The other parts of the xml structure _may_ follow.