updated SyntaxTagger
The syntax tagger is reading input from stdin and pipes it toward the berkeley parser. The input needs to be one sentence per line. An empty line closes the input stream and the program goes on processing the lines. Then it waits for a response by the berkeley parser and prints its stdout.
As it is now, the tagger does nothing more than the berkeley parser by itself besides providing some status information on stderr.